4.17: bloglist expansion, step5
Expand blogs so that each blog contains information on the creator of the blog.
Modify adding new blogs so that when a new blog is created, any user from the database is designated as its creator.

Modify listing all blogs so that the creator's user information is displayed with the blog.